leave out long

+++ src/main/java/atonkish/hexborder/integration/autoconfig/annotation/ModConfigEntry.java
@@ -9,6 +9,6 @@ public class ModConfigEntry {
     @Retention(RetentionPolicy.RUNTIME)
     @Target(ElementType.FIELD)
     public @interface HexagonSide {
-        long min();
+        int min();
     }
 }
